+++ docs/releases.rst
@@ -4,6 +4,7 @@ Release History
 0.1.1 - In development
 -------------------------
 
+* Added overall coverage display and tracking.
 
 0.1.0 - 24 September 2013
 -------------------------